fix: address merge review feedback
