+++ src/activation/repositoryMode.ts
@@ -279,7 +279,6 @@ export async function activateRepositoryMode(
                     await refreshService.refreshConflictUi();
                 },
             });
-            await moveUndockedEditorToNewWindow();
         } catch (error) {
             const message = getErrorMessage(error);
             showTimedWarningMessage(
